Add brain test to ensure that only uncached buildpacks are shipped
In order to do so a special inspector buildpack is added which inspects all
other buildpacks at staging time for cached dependencies.
An empty app is then pushed which won't be accepted by any of the regular
buildpacks. Only the inspector buildpack will accept the app if the "uncached"
check passes.